+++ .env.development
@@ -5,7 +5,7 @@ NODE_ENV = development
 VITE_TITLE = luckyColor-admin
 
 # 接口地址
-VITE_API_BASEURL = http://localhost:8081
+VITE_API_BASEURL = /api
 
 # 本地端口
 VITE_PORT = 9900
