+++ src/main.ts
@@ -19,8 +19,8 @@ const naive = create({
 });
 
 const app = createApp(App);
-app.use(router);
 app.use(store);
+app.use(router);
 app.use(naive);
 app.use(luckyColor);
 app.mount('#app');
