--- src/router/contentRouting.ts
@@ -1,9 +1,0 @@
-const routes = [
-  {
-    path: '/layouts',
-    name: 'layout',
-    component: () => import('@/layouts/normal/index.vue')
-  }
-];
-
-export default routes;
