+++ src/utils/message.ts
@@ -1,4 +1,4 @@
 import { createDiscreteApi } from 'naive-ui';
 
 // 为Naive UI组件创建离散的API
-export const { message, notification } = createDiscreteApi(['message', 'notification']);
+export const { message, notification } = createDiscreteApi([ 'message', 'notification' ]);
